Move FastInput component to src/components/elements/samples/
