+++ app/packs/src/components/elements/samples/FastInput.js
