+++ app/packs/src/components/elements/samples/SampleDetails.js
@@ -63,7 +63,7 @@ import NotificationActions from 'src/stores/alt/actions/NotificationActions';
 import MatrixCheck from 'src/components/common/MatrixCheck';
 import AttachmentFetcher from 'src/fetchers/AttachmentFetcher';
 import NmrSimTab from 'src/components/elements/samples/nmr_sim_tab/NmrSimTab';
-import FastInput from 'src/components/FastInput';
+import FastInput from 'src/components/elements/samples/FastInput';
 import ScifinderSearch from 'src/components/scifinder/ScifinderSearch';
 import ElementDetailSortTab from 'src/components/elements/ElementDetailSortTab';
 import { addSegmentTabs } from 'src/components/generic/SegmentDetails';
